Relocate EnableK8s impl
